Add tests

+++ ADO.NET/DAL/OrderDetails.cs
@@ -8,8 +8,8 @@
 
         public int ProductId { get; set; }
 
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
 
-        public int Quantity { get; set; }
+        public short Quantity { get; set; }
     }
 }
